+++ src/main/java/fuzs/betteranimationscollection/BetterAnimationsCollection.java
@@ -40,6 +40,7 @@ public class BetterAnimationsCollection {
     public static final AbstractElement HUMANOID_KNEES = REGISTRY.register("humanoid_knees", () -> new HumanoidKneesElement(), Dist.CLIENT);
     public static final AbstractElement ANIMATED_PLAYER = REGISTRY.register("animated_player", () -> new AnimatedPlayerElement(), Dist.CLIENT);
     public static final AbstractElement FAMILIAR_HORSE = REGISTRY.register("familiar_horse", () -> new FamiliarHorseElement(), Dist.CLIENT);
+    public static final AbstractElement SPITFUL_LLAMA = REGISTRY.register("spitful_llama", () -> new SpitfulLlamaElement(), Dist.CLIENT);
 
     public BetterAnimationsCollection() {